Changed class name to Events and moved method over.
